refactor PokemonFavoriteListViewModel
